fix: remove unused dependencies, add undeclared dependencies (#230)
* build: remove unused dependencies

The eslint config is broken anyway, so I also removed the reference to an
eslint dependency that wasn't declared as a dependency in the right package
and would have just happened to work due to hoisting.

* fix: declare all dependencies

These just happened to work due to dependency tree hoisting, but could have
easily broken in the future.

This should have been caught by eslint, but that's broken too.