fix(edge): note when stopped so we can kill process in early init
This is to address the edge case surfaced by the (Astro integration
PR)[https://github.com/withastro/astro/pull/13768] wherein the edge function
Deno server won't be closed if the .stop() comes in very soon after the init.

The issue is that .stop() can be called before denoBridge.runInBackground has
finished resolving (but before we've waited for the server to start), and in
that case we will never know it needed to be stopped.