ci: add separate CI workflow to fix node version issue
The new package has a min. of node 20.19.0 whereas the others have 20.6.1.

The CI setup in this monorepo assumes it can run `npm run test` on all workspace packages against
the earliest supported version (20.6.1), which is assumed to be consistent.

I considered several approaches here, and this seemed like the lesser of evils. It just sets the
`test` script in this package to a no-op (npm workspaces have no option to run something against
"all workspaces packages except this one") and added a separate CI workflow for the new package.
This also avoids another weirdness while I was at it, since this package has its own lint and
typechecking.