feat: add `@netlify/nuxt` Nuxt module (#344)
* feat: move req/res utils to @netlify/dev-utils

* build: add keywords to vite plugin package

* feat: add `@netlify/nuxt` Nuxt module

* refactor: remove useless directory?

* build: add a no-op build script

* fix: add missing deps and fix build steps

* fix: loosen nuxt module node reqs

* ci: add separate CI workflow to fix node version issue

The new package has a min. of node 20.19.0 whereas the others have 20.6.1.

The CI setup in this monorepo assumes it can run `npm run test` on all workspace packages against
the earliest supported version (20.6.1), which is assumed to be consistent.

I considered several approaches here, and this seemed like the lesser of evils. It just sets the
`test` script in this package to a no-op (npm workspaces have no option to run something against
"all workspaces packages except this one") and added a separate CI workflow for the new package.
This also avoids another weirdness while I was at it, since this package has its own lint and
typechecking.

* build: fix `test:types` nuxt module script

* test: add missing dependency to fixture

* chore(deps): remove unused dev dep

* fix(deps): specify a @types/node major

* chore: bump @netlify/nuxt version

I manually published 0.0.0 and 0.0.1 while testing