chore: fix outstanding eslint errors (#377)
* refactor: work around strange eslint rule bug

I did a bunch of research on this and I'm fairly confident it's a bug, but I couldn't come up with a
minimal repro to open an issue and felt like I couldn't justify spending more time here.

This is a valuable rule, so I'd rather not suppress it here or globally.

* refactor: remove unused eslint suppressions

---------

Co-authored-by: Eduardo Bouças <mail@eduardoboucas.com>