feat: add `@netlify/vite-plugin-tanstack-start`
This adds a framework-agnostic Vite build plugin that prepares a Vite app for deployment to Netlify. It expects that the
Vite app has an `ssr` environment (and no other server environments to deploy) that has exactly one bundle entry and
that this entry is a server request entry point, a module with a default export containing a Web Fetch handler under the
`fetch` property. TanStack Start meets all these conditions.

This build plugin can be opted into via the intentionally private, undocumented `build.enabled` option, passed into
`@netlify/vite-plugin`.

This also adds a new `@netlify/vite-plugin-tanstack-start` which just wraps `@netlify/vite-plugin` with this enabled.

At the moment, this only supports TanStack Start `alpha`, but this will be released imminently (presumably v.1.132 or
so). As such, it inherits its requirements:
- Vite 7+
- Node.js 22.12.0+